Modify: js 분리 및 수정
