+++ client/lib/dom/index.js
@@ -2,10 +2,10 @@
 
 export * from './css.js';
 export * from './attr.js';
-// export * from './swiper.js';
+export * from './swiper.js';
 export * from './insert.js';
 export * from './getNode.js';
-// export * from '../../js/partials/userList.js';
+export * from '../../js/partials/userList.js';
 export * from './showAlert.js';
 export * from './bindEvent.js';
 export * from './getInputValue.js';
